+++ services/eavesdropper/goproxy_eavesdropper.go
@@ -2,6 +2,7 @@ package eavesdropper
 
 import (
 	"bufio"
+	"fmt"
 	logger "log"
 	"net"
 	"net/http"
@@ -10,22 +11,31 @@ import (
 
 	"github.com/elazarl/goproxy"
 	"github.com/moooofly/goproxy/services"
+	"github.com/moooofly/goproxy/utils"
+	"github.com/moooofly/goproxy/utils/authx"
 )
 
+const realm = "EavesDropper-Realm"
+
 type EavesdropperArgs struct {
-	Local *string
+	Local    *string
+	White    *string
+	AuthFile *string
+	Auth     *[]string
 }
 
 type Eavesdropper struct {
-	cfg    EavesdropperArgs
-	log    *logger.Logger
-	isStop bool
+	cfg       EavesdropperArgs
+	basicAuth authx.BasicAuth
+	log       *logger.Logger
+	isStop    bool
 }
 
 func NewEavesdropper() services.Service {
 	return &Eavesdropper{
-		cfg:    EavesdropperArgs{},
-		isStop: false,
+		cfg:       EavesdropperArgs{},
+		basicAuth: authx.BasicAuth{},
+		isStop:    false,
 	}
 }
 
@@ -37,6 +47,7 @@ func (s *Eavesdropper) StopService() {
 		} else {
 			s.log.Printf("service eavesdropper proxy stopped")
 		}
+		s.basicAuth = authx.BasicAuth{}
 		s.cfg = EavesdropperArgs{}
 		s.log = nil
 		s = nil
@@ -48,18 +59,38 @@ func (s *Eavesdropper) Start(args interface{}, log *logger.Logger) (err error) {
 	s.log = log
 	s.cfg = args.(EavesdropperArgs)
 
+	err = s.InitBasicAuth()
+	if err != nil {
+		return
+	}
+
+	//s.basicAuth.Show()
+
+	wl, err := utils.LoadWhiteList(*s.cfg.White)
+	if err != nil {
+		return err
+	}
+
+	var conds []goproxy.ReqCondition
+	for _, v := range wl {
+		conds = append(conds, goproxy.Not(goproxy.ReqHostMatches(regexp.MustCompile(fmt.Sprintf("^.*%s.*$", v)))))
+	}
+
 	for _, addr := range strings.Split(*s.cfg.Local, ",") {
 		if addr != "" {
 
-			// ref: https://github.com/elazarl/goproxy/blob/master/examples/goproxy-eavesdropper/main.go
-
 			proxy := goproxy.NewProxyHttpServer()
-			proxy.OnRequest(goproxy.ReqHostMatches(regexp.MustCompile("^.*baidu.com$"))).
-				HandleConnect(goproxy.AlwaysReject)
+
+			// FIXME(mooofly): hardcode here right now
+			proxy.Verbose = true
+			proxy.Logger = s.log
+
+			proxy.OnRequest(conds...).HandleConnect(goproxy.AlwaysReject)
+
 			proxy.OnRequest(goproxy.ReqHostMatches(regexp.MustCompile("^.*$"))).
 				HandleConnect(goproxy.AlwaysMitm)
-				// enable curl -p for all hosts on port 80
-			proxy.OnRequest(goproxy.ReqHostMatches(regexp.MustCompile("^.*:80$"))).
+
+			proxy.OnRequest(goproxy.ReqHostMatches(regexp.MustCompile("^.*$"))).
 				HijackConnect(func(req *http.Request, client net.Conn, ctx *goproxy.ProxyCtx) {
 					defer func() {
 						if e := recover(); e != nil {
@@ -68,6 +99,13 @@ func (s *Eavesdropper) Start(args interface{}, log *logger.Logger) (err error) {
 						}
 						client.Close()
 					}()
+
+					ua := req.Header.Get("User-Agent")
+					s.log.Printf("User-Agent got: %s\n", ua)
+
+					ua2 := ctx.Req.Header.Get("User-Agent")
+					s.log.Printf("User-Agent2 got: %s\n", ua2)
+
 					clientBuf := bufio.NewReadWriter(bufio.NewReader(client), bufio.NewWriter(client))
 					remote, err := net.Dial("tcp", req.URL.Host)
 					if err != nil {
@@ -97,7 +135,7 @@ func (s *Eavesdropper) Start(args interface{}, log *logger.Logger) (err error) {
 						}
 					}
 				})
-			//proxy.Verbose = *verbose
+
 			err = http.ListenAndServe(addr, proxy)
 		}
 	}
@@ -106,4 +144,27 @@ func (s *Eavesdropper) Start(args interface{}, log *logger.Logger) (err error) {
 
 func (s *Eavesdropper) Clean() {
 	s.StopService()
+}
+
+func (s *Eavesdropper) InitBasicAuth() (err error) {
+	s.basicAuth = authx.NewBasicAuth(s.log)
+
+	if *s.cfg.AuthFile != "" {
+		n, err := s.basicAuth.AddFromFile(*s.cfg.AuthFile)
+		if err != nil {
+			return err
+		}
+		s.log.Printf("auth data added from file %d, total:%d", n, s.basicAuth.Total())
+	}
+
+	if len(*s.cfg.Auth) > 0 {
+		n := s.basicAuth.Add(*s.cfg.Auth)
+		s.log.Printf("auth data added from CLI %d, total:%d", n, s.basicAuth.Total())
+	}
+
+	return
+}
+
+func (s *Eavesdropper) Verify(user, passwd string) bool {
+	return s.basicAuth.Check(user, passwd)
 }