+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+<a name="0.3.0"></a>
+# [0.3.0](https://github.com/moooofly/goproxy/compare/v0.2.0...v0.3.0) (2019-02-18)
+
+
+### Features
+
+* support AES and get custom User-Info for later analysis ([00ef0a3](https://github.com/moooofly/goproxy/commit/00ef0a3))
+* update eavesdropper service as basic ([8bf38af](https://github.com/moooofly/goproxy/commit/8bf38af))
+
+
+
 <a name="0.2.0"></a>
 # [0.2.0](https://github.com/moooofly/goproxy/compare/v0.1.0...v0.2.0) (2019-02-13)